recommendation_engine/requirements.txt
pandas
scikit-learn
google-cloud-storage
google-cloud-bigquery
requests
pyyaml
